My favorite new feature for Symfony 2.8 is Guard. It makes creating custom and crazy authentication systems really really easy. I'm a bit biased: Guard was my creation, inspired by a lot of people and projects.
The Weirdest Login Form Ever

Later, I'll do some in-depth screencasts about Guard, but I want to give you a taste of what's possible. In this project, I have a tradition login system:

 42 lines  src/AppBundle/Controller/DefaultController.php
... lines 1 - 21
  /**
   * @Route('/login', name='login')
   */
  public function sillyLoginAction(){
    {dollar_sign}error = {dollar_sign}this->get('security.authentication_utils')
      ->getLastAuthenticationError();
    return {dollar_sign}this->render('default/login.html.twig', [
      'error' => {dollar_sign}error
    ]);
  }
... lines 34 - 42
The /login controller renders a login template and this template builds a form that POSTs right back to that same loginAction:

 32 lines  app/Resources/views/default/login.html.twig
... lines 1 - 9
  <form action='{{ path('login') }}' method='POST'>
... line 11
    <label for='exampleInputEmail1'>Username</label>
    <input type='text' name='_username' class='form-control' id='exampleInputEmail1' placeholder='Username'>
... lines 14 - 15
    <label for='exampleInputPassword1'>Password</label>
    <input type='password' name='_password' class='form-control'
      id='exampleInputPassword1' placeholder='Password'>
... lines 19 - 20
    <label for='the-answer'>The answer to life the universe and everything</label>
    <input type='text' name='the_answer' class='form-control' id='the-answer'>
... lines 23 - 25
     <input type='checkbox' name='terms'> I agree to want to be logged in
... lines 27 - 28
    <button type='submit' class='btn btn-default'>Login</button>
  </form>
... lines 31 - 32
It has a username field, a password field and then a couple of extra fields. One of them is 'the answer to life the universe and everything', which of course must be set to 42. The user also needs to check this 'agreement' checkbox.

So how can we make a login system that checks four fields instead of just the usual username and password? Hello Guard.

Before coding, start the built-in web server, which is now bin/console server:run:

bin/console server:run
Try out /login.