getCredentials()
Once we hook it up, the getCredentials() method will be called on the authenticator on every single request. This is your opportunity to collect the username, password, API token or whatever credentials the user is sending in the request.

For this system, we don't need to bother looking for the credentials unless the user is submitting the login form. Add an if statement: if {dollar_sign}request->getPathInfo() - that's the current, cleaned URL - != /login or ! {dollar_sign}request->isMethod('POST'), then return null.

When you return null from getCredentials(), no other methods will be called on the authenticator.

Below that, return an array that contains any credential information we need. Add a username key set to {dollar_sign}request->request->get('_username'): that's the name of the field in the form.

Repeat that for password and answer. In the form, answer's name is the_answer and the last is named terms. Finish the array by fetching the_answer and terms.

The keys on the right are obviously the names of the submitted form fields. The keys on the left can be anything: you'll see how to use them soon.