getUser()
If getCredentials() returns a non-null value, then getUser() is called next. I have a really simple User entity that has basically just one field: username.

I'm not even storing a password. Whatever your situation, just make sure you have a User class that implements UserInterface and a 'user provider' for it that's configured in security.yml:

 31 lines  app/config/security.yml
... lines 1 - 2
security:
... lines 4 - 5
  providers:
    my_entity_users:
      entity:
        class: AppBundle:User
... lines 10 - 31
See the  variable? That's equal to whatever you returned from getCredentials(). Set a new {dollar_sign}username variable by grabbing the username key from it.
Let's get crazy and say: 'Hey, if you have a username that starts with an @ symbol, that's not okay and you can't login'. To fail authentication, return null.

Now, query for the User! We need the entity manager, so add a private  property and generate the constructor. Type-hint it with EntityManager.

Back in getUser(), this is easy: return {dollar_sign}this->em->getRepository('AppBundle:User')->findOneBy() with username equals {dollar_sign}username.

The job of getUser() is to return a User object or null to fail authentication. It's perfect.