Handling Authentication Success and Failure
That's it! The last few methods handle what happens on authentication failure and success. Both should return a Response object, or null to do nothing and let the request continue.

onAuthenticationFailure()

The {dollar_sign}exception passed to onAuthenticationFailure holds details about what went wrong.

Was the user not found? Were the credentials wrong? Store this in the session so we can show it to the user. For the key, use the constant: Security::AUTHENTICATION_ERROR.

This is exactly what the normal form login system does: it adds an error to this same key on the session.

In the controller, the security.authentication.utils service reads this key from the session when you call getLastAuthenticationError():
lines  src/AppBundle/Controller/DefaultController.php
... lines 1 - 8
class DefaultController extends Controller {
... lines 11 - 21
  /**
   * @Route('/login', name='login')
   */
  public function sillyLoginAction(){
    {dollar_sign}error = {dollar_sign}this->get('security.authentication_utils')
      ->getLastAuthenticationError();
... lines 29 - 32
  }
... lines 34 - 40
}

Other than storing the error, what we really want to do when authentication fails is redirect back to the login page. To do this, add the Router as an argument to the constructor and use that to set a new property. I'll do that with a shortcut.

Now it's simple: {dollar_sign}url = {dollar_sign}this->router->generate() and pass it login - that's the route name to my login page. Then, return a new RedirectResponse

onAuthenticationSuccess()

When authentication works, keep it simple and redirect back to the homepage. In a real app, you'll want to redirect them back to the previous page.
There's a base class called AbstractFormLoginAuthenticator that can help with this.