start() - Inviting the User to Authentication
The start() method is called when the user tries to access a page that requires login as an anonymous user. For this situation, redirect them to the login page.

supportsRememberMe()

Finally, if you want to be able to support remember me functionality, return true from supportsRememberMe()
You'll still need to configure the remember_me key in the firewall.