How big is Symfony, really? It's HUGE! I'm kidding. Dude, Symfony is just a bunch of little libraries, so we can make it as small or as huge as we want. Most of us start with the Standard Edition: it looks basically like this, 10 or so configuration files, two front controllers, and some other skeleton files. Probably 20 or 30 files to begin with.
But no more in Symfony 2.8! Ok, the Standard Edition of course still exists, but now we have a brand new tool in the arsenal: the MicroKernelTrait. You can build a Symfony app that's as small as one file, then let it grow from there.

Hey, should we try it?

Basic 'Micro' Kernel

Imagine that we don't have this big project: we only have a composer.json file that requires symfony/symfony:

 64 lines  composer.json {
... lines 2 - 12
  'require': {
... line 14
    'symfony/symfony': '3.0.*@dev',
... lines 16 - 25
  },
... lines 27 - 62
}
In the app directory, create a new class called LittleKernel. The kernel is the heart of your application: make it extend the base Kernel class.

Normally, this class must implement two abstract methods: registerBundles() and registerContainerConfiguration(). But let's not implement those. Instead, use the new MicroKernelTrait. This implements registerContainerConfiguration() for us. All we need to do is add registerBundles(), configureRoutes() and configureContainer().

k, stop! Check this out, it's really cool. What is an application? Well, it's a set of bundles, a collection of services, and a list of routes. And would you look at that! We have exactly three methods that define the three things that make up a framework.