Single-File Symfony
Let's make a real-live app in just this one file. First, register some bundles
We only need one bundle: FrameworkBundle. To configure that bundle, head to configureContainer(), call {dollar_sign}c->loadFromExtension(), and pass it framework and an array of configuration. The only key it needs is secret. Set it to the top secret super cool-kids password: micr0

And guess what? We already know exactly how this loadFromExtension() works: it's the PHP equivalent of having a config.yml file with a framework key at the root and a secret key below that:

 61 lines  app/config/config.yml
... lines 1 - 10
framework:
... lines 12 - 13
  secret:          '%secret%'
... lines 15 - 61
Oh, and yea, you can still load YML and XML config files, and you probably will. I'll mention that soon.

At this point, this app will already work. But let's make a page. In configureRoutes(), we receive a shiny new RouteCollectionBuilder object - another new thing for Symfony 2.8 that makes adding routes in PHP a lot more fun than it used to be.

Add a route with {dollar_sign}routes->add('/hello/symfony/{version}'). The second argument is the controller: pass it kernel:helloSymfony.

This is the service format for configuring a controller: kernel is the name of the service - that actually points to this object - and helloSymfony is the name of the method.

Hey, let's build that: public function helloSymfony(). Give it a  argument and say return new Response('Hi Symfony version '.{dollar_sign}version)

OMG. That's a fully-functional Symfony app in one file. The only thing we're missing is a front controller that boots and runs this guy. To be really trendy, we could put that code at the bottom of this file. But come on guys - I think we're letting this micro-framework fad get to our heads. Some structure is a good thing.