Adding the Front Controller
Instead, in the web directory, create a new file called tiny.php. This will be our version of an app.php or app_dev.php file.

Start by requiring Composer's autoloader and our LittleKernel.php, since it isn't configured to be autoloaded.

The rest of this file is just the same boring stuff we see in app_dev.php. In fact, copy the bottom of that file and paste it here. Change AppKernel to LittleKernel and make sure the Request class has its use statement. Remove loadClassCache() - a small performance line - to make this even smaller

We're crazy!

Ready? Try it!

Over in the browser add /tiny.php to the URL:

http://localhost:8000/tiny.php
Ah! Hide from the error! Wait, check it out, it says:

No route found for 'GET /' in...
Our app is working! We're using such a small bit of Symfony that we don't even have the nice exception pages. You can get these back by adding TwigBundle, but it's not technically necessary.

Try the real page now: /hello/symfony/3. There it is!