Creating a Bigger (Micro) App
A real app won't be just one file. But the MicroKernelTrait allows you to grow and opt into whatever features you want. Let's see if we can get our existing app running. That means loading annotation routes from DefaultController and booting Twig to render the templates.

In LittleKernel, add TwigBundle and SensioFrameworkExtraBundle to registerBundles(). To activate Twig, we need more configuration under the framework key. Add a templating key with an engines sub-key. In there, add twig in an array.
That configuration looks a little 'deep', but it's exactly what you've always had in your config.yml file:

 61 lines  app/config/config.yml
... lines 1 - 10
framework:
... lines 12 - 21
  templating:
    engines: ['twig']
... lines 24 - 61
Also activate the assets subsystem with an assets key set to an empty array. This makes it possible to use the asset() function in Twig