Loading External Routing Files
Next, we need to load the annotation routes. That's super easy.

Use {dollar_sign}routes->import(). Pass it whatever routing resource you want, like a YAML or XML file. In our case, import the annotations from the Controller directory with __DIR__.'/../src/AppBundle/Controller'. Pass / as the second argument - that's the prefix - and annotation - the 'type' - as the last arg

This should feel familiar too: it's equivalent to importing routes in a YAML file with resource and type keys:

4 lines  app/config/routing.yml
app:
  resource: '@AppBundle/Controller/'
  type:     annotation
Really it's the exact same thing, just done in PHP instead.

We're done! But please, please don't get too excited, it's not going to work quite yet. Refresh anyways to see what's happening.

Annotations Loader

Error!

The annotation '@Sensio\Bundle\FrameworkExtraBundle\Configuration\Route' in method could not be auto-loaded.