If you work with annotations, then you need one little extra line of code. This line already exists inside of Symfony's normal app/autoload.php file:
16 lines  app/autoload.php
... lines 1 - 2
use Doctrine\Common\Annotations\AnnotationRegistry;
... lines 4 - 12
  AnnotationRegistry::registerLoader(array({dollar_sign}loader, 'loadClass'));
  ... lines 14 - 16
In tiny.php, require this autoload file. That gives us the annotation line we need and still initializes Composer's autoloader.

Refresh! It's alive!

Now check out the homepage. It's dead again!

Unknown 'is_granted' function in base.html.twig at line 19.
Open that file and find line 19. Our app is choking on is_granted():

 39 lines  app/Resources/views/base.html.twig
... lines 1 - 16
  <ul class='nav nav-pills pull-right'>
    <li role='presentation' class='active'><a href='{{ path('homepage') }}'>Home</a></li>
     {% if is_granted('IS_AUTHENTICATED_FULLY') %}
       <li role='presentation'><a href='{{ path('logout') }}'>Logout</a></li>
     {% else %}
       <li role='presentation'><a href='{{ path('login') }}'>Login</a></li>
     {% endif %}
   </ul>
... lines 25 - 39
This is one of the best and worst parts of the MicroKernel. If you're accustomed to having twenty bundles, then you're accustomed to having all the features of Symfony. But, with the MicroKernel you must opt into each feature. The is_granted() comes from Symfony's SecurityBundle... and we're not using that! If we need it, then you'll need to add it to LittleKernel and configure your security.yml file.