Opting Into Features
Open that file and find line 19. Our app is choking on is_granted():

 39 lines  app/Resources/views/base.html.twig
... lines 1 - 16
  <ul class='nav nav-pills pull-right'>
    <li role='presentation' class='active'><a href='{{ path('homepage') }}'>Home</a></li>
    {% if is_granted('IS_AUTHENTICATED_FULLY') %}
      <li role='presentation'><a href='{{ path('logout') }}'>Logout</a></li>
    {% else %}
      <li role='presentation'><a href='{{ path('login') }}'>Login</a></li>
    {% endif %}
  </ul>
... lines 25 - 39
This is one of the best and worst parts of the MicroKernel. If you're accustomed to having twenty bundles, then you're accustomed to having all the features of Symfony. But, with the MicroKernel you must opt into each feature. The is_granted() comes from Symfony's SecurityBundle... and we're not using that! If we need it, then you'll need to add it to LittleKernel and configure your security.yml file.

For now, remove is_granted()

And try again.

It alive... again! That's the MicroKernelTrait in a nut shell: add as many bundles as you want, configure them and add routing. You can still load external routing and configuration files, and you should! The point isn't to put everything into a single file: it's about starting with a single file and then choosing which bundles and configurations make sense for your project.

Right now, I have my framework configuration in PHP. But if the project grows, I might want to use a YAML file instead. That's no problem.

 48 lines  app/LittleKernel.php
... lines 1 - 38
    protected function configureContainer(ContainerBuilder {dollar_sign}c, LoaderInterface {dollar_sign}loader)
... lines 40 - 48
See that  variable? It has an import() method on it - use that to import a config.yml file and you're done. This is the PHP-equivalent to the imports line in YAML.