PhpStorm gives me a nice skeleton for the class.... but now it has too much. Remove the getName() method at the bottom. This was always a useless, but required function. But no more!
For the most part, forms look the same. In configureOptions() call {dollar_sign}resolver->setDefaults() and pass it data_class set to AppBundle\Entity\User

My super-simple user has just one field on it: username, which makes this one of the most ridiculously useless registration forms in history:
56 lines  src/AppBundle/Entity/User.php
... lines 1 - 11
class User implements UserInterface {
... lines 14 - 20
  /**
   * @ORM\Column(type='string')
   */
  private {dollar_sign}username;
... lines 25 - 30
  public function getUsername() {
    return {dollar_sign}this->username;
  }
... lines 35 - 50
  public function setUsername({dollar_sign}username) {
   {dollar_sign}this->username = {dollar_sign}username;
  }
}
Form types as Class Names!

Add our one field with {dollar_sign}builder->add('username'). But stop! Here is where things are different. Before I would have passed text as the second argument. But now, I'm going pass the full class name to the class that's behind the text field type.

To make things easier, use a shortcut: TextType::class

PhpStorm doesn't auto-complete classes when you do this... yet. But if you focus on the class, a light-bulb will allow you to import the class. Or hit option+enter on a Mac to bring it up immediately. This adds the use statement. Using ::class is new in PHP 5.5... so this change won't be much fun unless you can use this. But, you can keep using the old syntax until you switch to 3.0... which requires 5.5 anyways.

Phew! That's big thing number one: instead of weird strings, we have full class names. This sucks because... well, this is a little bit more typing. But this is sweet because a class name is more meaningful than a string like text. If this was your first time using Symfony, you'd have a better chance of understanding how things work.