In app/Resources/views/default, create that register.html.twig template. I'll copy in some boilerplate code for this form. None of this has changed. This opens the form tag, dumps out all the fields, adds a submit button and closes the form.
Time to test it! Head to /register, throw in our username: leannapelham and hit register. There's our dump!

Here's the moral of the story: forms work exactly like before. But now, form classes are always referred to by their class name: never objects and never the short, but weird magic string from the past.

This might be the biggest change in Symfony 3 that people will complain about, and rightly so: if you have a lot of forms, this is a lot of changes. For that, check out the Symfony Upgrade Fixer(https://github.com/umpirsky/Symfony-Upgrade-Fixer): a library that can automate some of the tasks of upgrading, including this one.