We all love ice cream and Symfony's console. We love the ability to add colors, beer shaped progress bars(http://knpuniversity.com/blog/fun-with-symfonys-console), tables and a bunch of other cool stuff.
A little feature snuck into Symfony 2.7 and was improved greatly in Symfony 2.8. It's called the SymfonyStyle. It's more than just a cool way to dress, it's the Twitter Bootstrap for styling console commands. Ok ok, it's not a huge deal, but let's face it, that blog post with the beer icon in the console? It's pretty much our most popular blog post... so clearly you guys love this stuff.

Create a new Command directory. Inside, I'll take the lazy road and have PhpStorm make me a new command called StylesPlayCommand. Give it a name! styles:play.

Start like we always do, with lame, worn-out ->writeln('boring'). Zip over to the terminal and try that out:

./bin/console styles:play
And there it is! In all its white text on a black background glory.