Create the Voter
In the Security directory, create a new class called UserVoter and make this extend Voter.
This is a new class in Symfony 2.8 that's easier than the old AbstractVoter.

Use command+n to open the generate menu and select 'Implement methods'. The two methods you need are supports and voteOnAttribute.
This is a little different than before.

Now stop! And go register this as a service. Call it user_voter and add its class: UserVoter. There aren't any arguments, but you do need a tag called security.voter

As soon as we give it this tag, the supports method will be called every time we call isGranted(), asking our voter 'Yo! Do you support this attribute, like ROLE_USER or USER_VIEW?'.

I'm already logged in - so I'll head to /users/weaverryan. Access denied! Right now, none of the voters are voting on this: they are all saying that they don't support the USER_VIEW attribute. If nobody votes, access is denied.